requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.6.3
astropy==5.2.2
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.8.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cycler==0.12.1
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.6.1
email-validator==2.2.0
entrypoints==0.3
ephem==4.1.5
exceptiongroup==1.2.1
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
geographiclib==2.0
geopy==2.4.1
gradio==4.38.1
gradio-client==1.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.14.0
httpx==0.27.0
huggingface-hub==0.23.5
hyperlink==19.0.0
idna==2.8
importlib-metadata==8.0.0
importlib-resources==6.4.0
incremental==16.10.1
itsdangerous==2.2.0
jinja2==3.1.4
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
orjson==3.10.6
packaging==24.1
pandas==2.0.3
pexpect==4.6.0
pillow==10.3.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.8.2
pydantic-core==2.20.1
pydub==0.25.1
pyephem==9.99
pyerfa==2.0.0.3
pygments==2.18.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python-multipart==0.0.9
python-telegram-bot==13.3
pytz==2024.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.7.1
ruff==0.5.2
SecretStorage==2.3.1
semantic-version==2.10.0
service-identity==18.1.0
shellingham==1.5.4
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sos==4.4
ssh-import-id==5.10
starlette==0.37.2
systemd-python==234
telegram==0.0.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
Twisted==18.9.0
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.2
uvicorn==0.30.1
wadllib==1.3.3
websockets==11.0.3
werkzeug==3.0.3
wifite==2.5.2
zipp==3.19.2
zope.interface==4.7.1
